requirements.txt
asgiref==3.7.2
Django==4.1
django-taggit==3.0.0
sqlparse==0.4.4
psycopg2-binary==2.9.3